+++ src/App.js
@@ -4,7 +4,6 @@ import { BreakpointProvider, setDefaultBreakpoints } from 'react-socks';
 import HorNav from './components/Navigation/Navigation';
 import LandingPage from './views/LandingPage';
 import Dashboard from './views/Dashboard';
-import SignUp from './views/SignUp';
 import Landlord from './views/Landlord';
 import Tenant from './views/Tenant';
 import Features from './views/Features';
@@ -20,7 +19,6 @@ const App = () => {
         <Router>
           <LandingPage path="/" />
           <Dashboard path="/dash" />
-          <SignUp path="/signup" />
           <Landlord path="/landlord" />
           <Tenant path="/tenant" />
           <Features path="/features" />
