+++ src/components/Auth/AuthFlip.js
@@ -6,47 +6,44 @@ import SignUp from './SignUpForm/SignUp';
 import { AuthFlipContainer } from '../UI';
 
 const AuthFlip = ({ loginSubmit, signupFn, show, setShow }) => {
-  const [flipped, setFlip] = useState(false);
+  const [flipped, setFlip] = useState(true);
 
   const toggle = () => {
     return setFlip(!flipped);
   };
 
   const { transform, opacity } = useSpring({
-    opacity: flipped ? 1 : 0,
+    opacity: [1, 0, 1],
     transform: `perspective(600px) rotateY(${flipped ? 180 : 0}deg)`,
     config: { mass: 5, tension: 500, friction: 80 }
   });
 
   const AuthFlipCard = a(AuthFlipContainer);
-  const AuthFlipBack = a(AuthFlipContainer);
 
   return (
     <>
       <AuthFlipCard
-        style={{ opacity: opacity.interpolate(o => 1 - o), transform }}
-      >
-        <LoginForm
-          submit={loginSubmit}
-          toggleFlip={toggle}
-          show={show}
-          setShow={setShow}
-        />
-      </AuthFlipCard>
-
-      <AuthFlipBack
         style={{
           opacity,
-          transform: transform.interpolate(t => `${t} rotateY(180deg)`)
+          transform
         }}
       >
-        <SignUp
-          submit={signupFn}
-          toggleFlip={toggle}
-          show={show}
-          setShow={setShow}
-        />
-      </AuthFlipBack>
+        {flipped ? (
+          <LoginForm
+            submit={loginSubmit}
+            toggleFlip={toggle}
+            show={show}
+            setShow={setShow}
+          />
+        ) : (
+          <SignUp
+            submit={signupFn}
+            toggleFlip={toggle}
+            show={show}
+            setShow={setShow}
+          />
+        )}
+      </AuthFlipCard>
     </>
   );
 };
